26-08-2023 002 Cambios agregados - se agrega Screen deslizable, se dejará en Stand-by

+++ lib/domain/entities/product_model_post.dart
@@ -4,7 +4,7 @@ class ProductModel {
   String? franquicia;
   int? volumen;
   String? descripcion;
-  double? precio;
+  int? precio;
   String? path;
 
   ProductModel({
